ML_API/requirements.txt
fastapi
numpy
uvicorn
Pillow
tensorflow-cpu
